+++ src/data/json_emitter.h
@@ -22,6 +22,7 @@
 
 #pragma once
 
+#include <functional>
 #include <string>
 #include "data/datastore.h"
 #include "data/record_dispatcher.h"
@@ -29,13 +30,21 @@
 class JsonRecordEmitter
 {
 public:
-    JsonRecordEmitter(const std::string& ui_dir, RecordDispatcher& dispatcher, DataStore& store);
+    using WaveStateLoadPolicy = std::function<bool(const DataStore&)>;
+
+    JsonRecordEmitter(
+        const std::string& ui_dir,
+        RecordDispatcher& dispatcher,
+        DataStore& store,
+        WaveStateLoadPolicy wave_state_load_policy = {}
+    );
     void run();
     void runOccupancyOnlyForTests();
 
 private:
     void emitMetadata();
     void emitWaveHierarchy();
+    void emitWaveStates();
     void emitOccupancy();
     void emitCounters();
     void emitRealtime();
@@ -48,4 +57,5 @@ private:
     std::string ui_dir;
     RecordDispatcher& dispatcher;
     DataStore& store;
+    WaveStateLoadPolicy wave_state_load_policy;
 };