+++ src/util/accordionwidget.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <QMouseEvent>
+#include <QMetaObject>
 #include <QPushButton>
 #include <QScrollArea>
 #include <QVBoxLayout>
@@ -31,10 +32,13 @@ signals:
     void expansionChanged(bool isExpanded); // Signal for external sync
 
 private:
+    void watchContentDestruction();
+
     QString m_title;
     bool m_isExpanded;
 
     QWidget* m_contentWidget;
+    QMetaObject::Connection m_contentDestroyedConnection;
     QVBoxLayout* m_internalLayout;
 
     int m_lastExpandedHeight;